requirements.txt
# Crypto AI Trading Bot - Requirements
# Python 3.11+

# PyBroker Framework
pybroker>=1.17.0

# Exchange & Data
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.32

# Technical Analysis
ta-lib>=0.4.28  # Optional - requires system library
pandas-ta>=0.3.14b0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Logging & Monitoring
loguru>=0.7.0
python-dotenv>=1.0.0

# Async & HTTP
aiohttp>=3.9.0

# Database (optional)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type hints
typing-extensions>=4.8.0
